Automatic queue file reset
Added a setting to reset the queue log file after X number of lines.
This only triggers when the queue is quiet. It safely truncates the
file in place and then performs a full sync in case inotify added to
the queue between the check and the reset. A full reset every now and
then seems like a good idea anyway.

It's not possible to run a sed command to delete only the lines read so
far because --in-place actually swaps out the file which messes up the
inotify writing process (understandably). I tested this extensively in
various ways and found queued entries were lost if the queue file was
altered outside of the process. Locks didn't help either (perhaps
because of the way IO is buffered) and they slowed things down
dramatically.