Added back close_write to notifications
MODIFY is not ideal as file could be written for an extended period of time (with multiple save actions but no closing) and every save action will be transmitted thus making an incomplete file to be sent.

CREATE (although documentation does not state this) - does send notifications not only on creation but on file writes as well. And is covering what modify providing.

WRITE_CLOSE - is a sure signal a file was finished writing (or at least exited write mode).

CREATE+CLOSE_WRITE seem to cover file writes and creations (file and folder), while MOVE, DELETE, ATTRIB cover file and folder remaining actions.