Added MODIFY to monitored actions
Based on tests and manual - it is possible to open a file for writing and never close it (keep appending to it) and if attributes are not changed - then it is possible that such file remains unnoticed. Only MODIFY notification provides notifications on files which are continuously being appended and never close write mode and never change file attribute (for example keeps same timestamps and permissions). rare but happens to miss. There seem to be  no harm performance wise to send file as it is being written to.