Recursive csync checking
This will ensure all files are checked by csync when inotify misses
files in nested dirs.

With this csync recursive mode there is no need to pass in files within a directory so in this current form there will be some duplicated checks. I shall strip out nested files in the next version.