Nested file filtering
Filtering out filenames in subdirectories as they're no longer required
now the csync update is recursive.

This is solid code and worked well. However, there was no worthwhile
speed increase in the csync check when passing one pathname vs. all the
files within. Better to keep the simplicity of the main branch.