Separate scripts for running PostgreSQL parallel and SQLite
PostgreSQL can help parallelizing submission of files to peers thus speeding up file transfers and deletions but the tests show that less than 4-5 nodes are slower/same speed as SQLite, but large number - over 10 peers could be faster as PostreSQL has no locking issues on deletions.

For this reason PostgreSQL solution is left for reference or if someone has a need to use with large number of peers.